+++ glossary.js
@@ -45,13 +45,15 @@ function glossaryItemFromDivElement(glossaryItemDivElement) {
     const ddElements = Array.prototype.slice.apply(glossaryItemDivElement.querySelectorAll('dd'));
     const relatedTermDdElements = ddElements.filter(ddElement => ddElement.getAttribute('class') === 'related-terms');
     const relatedTerms = (relatedTermDdElements.length > 0) ? glossaryItemRelatedTermFromDdElement(relatedTermDdElements[0]) : [];
+    const needsUpdatingDdElements = ddElements.filter(ddElement => ddElement.getAttribute('class') === 'needs-updating');
 
     return {
         terms: dtElements.map(glossaryItemTermFromDtElement),
         details: ddElements
-            .filter(ddElement => ddElement.getAttribute('class') !== 'related-terms')
+            .filter(ddElement => ddElement.getAttribute('class') !== 'related-terms' && ddElement.getAttribute('class') !== 'needs-updating')
             .map(ddElement => ddElement.textContent.trim()),
-        relatedTerms: relatedTerms
+        relatedTerms: relatedTerms,
+        needsUpdating: needsUpdatingDdElements.length > 0
     }
 }